+++ qcal/_vendor/__init__.py
@@ -1,0 +1,1 @@
+# Vendored third-party dependencies used internally by qcal